Merge into trunk all the work done by Min (Benjamin Ragan Kelley) on the weave cleanup branch.
Revisions merged: 3652:3734.

Note that there is still a wx test failing, and the weave.test(10) behavior
is still not fully correct, as it seems to be compiling some files in
directories it shouldn't.  But considering how when Min started there were
upwards of 30 failures of all kinds, the situation is much improved.

After Matthew commits all the testing updates to make it easier to write
more tests, the weave cleanup work should be continued.