Fix #623: better Cephes jv.c:recur continued fraction termination conditions
There were two problems in the old termination conditions for the
continued fraction for Jn/Jn-1:

1) It could terminate spuriously on the first iteration, because the
   initial value for `ans` corresponded to a possible value of the fraction
   but not one obtained from the iteration.

2) `jv` could call `recur` with values of `n` and `x` such that the
   continued fraction iteration would not converge and the maximum
   iteration limit was reached.

The code is now more careful with 1) and enforces a minimum number of
iterations, and 2) is worked around by bumping the maximum iteration
limit so that the CF should converge for all calls that come from `jv`.
The iteration limits can be estimated as the number of iterations
needed is known for this continued fraction.

Additional tests included.