ENH: optimize: refactor line searches to scalar search + line wrapper; move to linesearch.py
Each of the existing line search functions

    f(x + s*p) -> suitable minimum

is split to a scalar search function

    phi(s) -> suitable minimum

and a wrapper

    phi_p(s) = f(x + s*p)

that makes it a line search.

This refactoring makes the logic in the functions slightly clearer, and
makes reusing them in large-scale non-linear equation solvers (which
only need 1-d searches) more straightforward.

New tests for line searches are also added. These basically check that
they do what they promise, for a couple of problems.