BUG: spatial: ensure that qhull._find_simplex_directed always terminates
If the triangulation contains very narrow simplices, this may introduce
rounding error in the barycentric transformations, making the algorithm
think that two triangles are mutually closer to the target point. This
causes an infinite loop.

The solution is to limit the iteration count, and fall back to brute
force if the algorithm fails. More sophisticated cycle detection might
also be made --- however, the brute force O(N) cost probably overwhelms
any gains from such approaches, and adding cycle detection would
slow down the fast path, so it's probably best to keep it simple.