TEST: demonstrate some details. DONT PULL THIS
This commit replaces scipy.optimize.leastsq from
minpack.leastsq to lmfit.leastsq. This is a demonstration that
minpack.leastsq and lmfit.leastsq are actually equivalent.
lmfit.leastsq passes the minpack tests, but only by re-writing
ValueErrors into TypeErrors, which I consider ugly.

This is just a nice demonstration and should not
be pulled into mainline scipy. Maybe in the far future,
when everything is really well tested.