ENH: optimize: add _minimize_slsqp function with standard signature
As for COBYLA, constraints definition has been reviewed. They should now be
defined as a sequence of dictionaries with fields 'type', 'fun' and
optionally 'jac' and 'args' (although `fmin_slsqp` does not support
constraints extra arguments distinct from those of the objective function).

The constraints handling is similar to that of COBYLA. A large part of the
code has also been review (simplified in some respect) accordingly. In
particular, whether constraints are provided as a list of scalar functions
or as one vector function is transparent. Also, a lot of function wrapping
has been dropped.

Finally, an example has been added in `__main__` which tests both
`fmin_slsqp` and `_minimize_slsqp` with bounds and constraints.

`fmin_slsqp` now calls `_minimize_slsqp`, so backwards compatibility should
be preserved.