ENH: add constrained optimization solvers to the minimize interface
New methods L-BFGS-B, TNC, COBYLA and SLSQP in `minimize` for constrained
minimization. Additional parameters are `bounds` and `constraints`. As
described in the previous commits, `constraints` is either a dict or a
sequence of dict if several constraints are to be handled.

Check if provided parameters are relevant with respect to the choosen
methods. Returns warning if not.

`show_minimize_options` has been updated accordingly.

Tests for `minimize` with these new methods have been added.