BUG: spatial/qhull: discard degenerate barycentric transforms (fixes #1605)
For some point sets, the Delaunay triangulation produces nearly
degenerate simplices. These have singular barycentric transforms.

Previously, we discarded exactly singular transforms.  However,
computing the transform may succeed, but the result may still be invalid
and cause problems later on e.g. in _find_simplex_directed.

This commit adds a better check for validity of the transform: we
explicitly check the condition number, and drop those for which it is
large compared to machine precision.