Revert "TEST: demonstrate some details. DONT PULL THIS"
This reverts commit fe4432f6d6522afde18a7a2b540f9da4d3e688c4.

This patch had replaced the minpack leastsq with the lmfit
leastsq. This was intended for testing purposes only, so
that one can proove that indeed both functions work the same
and give the same result. As people started complaining
I wanted to kill good ole minpack leastsq (which I didn't
intend) I hereby revert the commit.