ENH: 'order' keyword for toarray() on sparse types
This adds an 'order' keyword to select the memory layout of the
resulting array when using toarray() to create a dense array
from a sparse matrix type. The default remains a C-contiguous
array. Tests are also updated to test both the default and
explicit 'C' and 'F' modes.

Generated output of SWIG will follow in a separate commit, so
that this one can be easily cherry-picked if what I generated
is unsatisfactory.