ENH: 'order' keyword for todense() on sparse types
This adds an 'order' keyword to select the memory layout of the
resulting object when using todense() to create a numpy matrix
object from a sparse matrix type, similarly to the previous
change to toarray(). The default remains a C-contiguous
matrix. Tests are also updated to test both the default and
explicit 'C' and 'F' modes.

Generated output of SWIG will follow in a separate commit, so
that this one can be easily cherry-picked if what I generated
is unsatisfactory.