Added method for t-test with unequal variances. ttest_ind assumes population variances are equal.
Refactor common code in all 4 t-test methods to _ttest_ind_setup and
_ttest_finish.

Fix bug in all 4 methods. Previous implementation would, if the
difference in means and the denominator of the t-statistic were zero,
set the t-statistic equal to 1. If the difference in means is zero, the
t-statistic should always be zero. Code and respective tests changed to
reflect this.

Added regression tests and tests that compares output to results of R
t.test() calls.