ENH: Added an out= argument to sparse toarray().
Mutually exclusive to the previously added 'order=' argument, this
should serve the same crowd, i.e. people who occasionally need a dense
copy of their sparse matrix but *really* can't afford to be allocating
an array that size over and over again. Refactored a bit into the base
class so as to not duplicate code between lil.py and coo.py.

Tests included.