TST: fix linalg.norm test failure.  Thanks to Rudiger Kessel.
The intended output was obviously 0.5 (where numpy.linalg.norm gives 0.0); this
seems to be the case on most but not all systems.  The reason is that snrm
seems to internally use double precision for most BLAS versions, but not all.

Note that test_overflow() for this implementation of norm() passes on all
systems, so even where test_stable() fails the performance of norm() looks
better than for numpy.linalg.norm().