BUG: process pyx files with Cython 0.17.1
This fixes import errors in Python 3.3, because Cython < 0.17 generated
C code that results in the python C import function being passed a
``level` of '-1'. The value of -1 meant (before python 3.3) that the
import can be absolute or relative. In Python 3.3, a level of -1 raises
an error:

   ValueError: level must be >= 0

(see python issue #15610).  Cython fixed this as of commit cb40a3e6264b,
released in Cython 0.17 (see output of ``git describe --contains
cb40a3e6264b``).