BUG: special: yet another fix to error propagation
The problem is that sf_error may be called somewhere deep in a special
function implementation, and this may raise a Python exception,
depending on the warnings settings.

If sf_error is defined for Cython as except *, Cython tends to swallow
these exceptions. The alternatives are then either to declare everything
as except *, or nothing as except * --- that is, either to check
PyErr_Occurred() everywhere, or check it nowhere (except after the ufunc
loop exits, Numpy checks it anyway).

For performance reasons (avoid unnecessary synchronization points due to
GIL locking), I'll pick here the latter option --- we raise the
exception without telling Cython about it, and rely on the usual ufunc
machinery collecting it.