BUG: special: fix build failure on mingw32
The issue is that Cython (0.17 at least) assumes that functions marked
"cdef extern" come from external DLL libraries, and appends
corresponding declaration attributes. However, this is incorrect in the
case of functions in scipy.special. The solution here was to move the
declarations to a separate .h file, outside of Cython's control.