FIX: Do not rely on strides for contiguous arrays in signal.medfilt
When an array is contiguous in memory but has shape[dim] == 1, then
its strides[dim] is not really used, so it can be considered arbitrary
even if the array is contiguous. This is generally true for 0-sized arrays.

Even without a change in numpy, this code (at least with numpy <1.8.)
could segfault for one dimensional arrays with one element.