Fixed a bug in scipy.stats.mstats.kurtosis() that caused a ValueError to be thrown when bias=False. In the same context, fixed a TypeError that was being thrown for one-dimensional input where all elements are masked.
Appropriate tests have been added to the testing suite for the function.

Fixed a bug in scipy.stats.mstats.kurtosis() that caused a ValueError to be
thrown when bias=False. In the same context, fixed a TypeError that was being
thrown for one-dimensional input where all elements are masked.

Appropriate tests have been added to the testing suite for the function.

Relates to ticket: http://projects.scipy.org/scipy/ticket/1798