BLD: Enforce Fortran ABI compatibility both for LAPACK and BLAS
Currently, Scipy contains some ABI compatibility wrappers for a few of
the BLAS routines, and a few LAPACK routines. Nevertheless, libraries in
Scipy (e.g. ARPACK) use parts of LAPACK and BLAS not covered.

Moreover, especially on OSX it can be easy for the user to mistakenly
link Accelerate (g77 ABI) with an ABI-incompatible LAPACK library.

For these reasons, it is better to be strict with the ABI compatibility
for the moment.