Update MAINTAINERS.rst.txt links on modules issues
Following this idea:

http://mail.scipy.org/pipermail/scipy-dev/2013-August/019040.html

I updtated the links so they point to the new GitHub issues tracker, instead of the old Trac.

I could not replace the RSS feed links because I could not find a direct equivalent on GitHub. I don't know if keeping them it's worth it - maybe the sooner all the references to Trac disappear, the better.