ENH qhull: raise a ValueError if masked array given as `points` argument
The `points` argument should contain the points coordinates. This has
no sense in the general case using masked array. In fact points could
not be partially defined, e.g. (X, --, Z) is not a point.

Moreover, the `points` array is always converted into a contiguous
array which does not contain the mask anymore. So, computations are
performed on the whole points set, even hidden points.

Closes gh-2675