add derivatives to bivariate splines
These changes add a fortran routine to return derivatives at arbitrary
points and includes derivative evaluation for Bivariate splines. Keyword
'grid' is added to the Bivariate spline evaluation to remove the
necessity of the 'ev' method for arbitrary point evaluation. These
changes have been tested as working with the 0.12 branch.