Guarantee oriented simplices for 2D delaunay
This commit updates the existing Delaunay code in order to ensure
that 2D triangulations maintain a consistent ordering. This
is similar to specifying the 'i' flag in QHull. This brings
the behaviour inline with Matlab, and is the most sensible output
for 2D triangulations.

This is achieved by swapping the first and second indices if
the orientation of the facet is clockwise. The logic is written
to try and minimise the impact on higher order triangulations.