ENH: signal: Allow upsweeps of frequency in the 'hyperbolic' chirp.
The restriction that f0 > f1 > 0 has been removed.  The only requirement
is that f0 and f1 are not zero.  Negative frequencies can be given.
(Even f0 and f1 with opposite signs are allowed, but this is not expected
to be a useful case.)  The degenerate case with f0 == f1 is now also
handled.

Two small changes not directly related to the above change:
* Changed "geometric" to "logarithmic" in the argument of the ValueError
  that can be raised when method="logaithmic".
* Added `t = asarray(t)` in the function _chirp_phase, so the 't' argument
  of `chirp` is now `array_like`.

Closes gh-3285.