MAINT: io: Remove empty strings warning from idl.py
This warning has been in the code for quite a while (therefore people who
use it already know about it), and it can be rather annoying and verbose
when reading large arrays with empty strings because it is printed for each
occurrence. I propose to get rid of these warnings altogether.