TST: integrate: Remove an invalid test of odeint.
In a test of the use of the banded jacobian arguments
of scipy.integrate.odeint, there is a check that two
different calls to odeint result in the same number
of jacobian evaluations at each step.  In one call,
the banded jacobian arguments are not used.  This is
an overly optimistic test; we can't really be sure that
the number of jacobian calls will be the same in these
two cases.