MAINT/DOC: optimize: Fix a doc build warning caused by optimize.anneal.
The function 'anneal' is decorated with '@deprecate(message="...")'.
When given a 'message' argument, 'numpy.deprecate' inserts
two lines at the beginning of the docstring.  The first is simply
'anneal is deprecated!'  The second is the message.  The 'deprecate'
decorator does not add any extra spaces to the beginning of these
lines, so the docstring ends up with its first two lines being
unindented.  This breaks the convention used in the rest of the
docstring, and confuses Sphinx.

This change is a quick fix: add four spaces to the beginning of
the 'message' argument.  Then only the first inserted line is
unindented, which Sphinx handles correctly.