BUG: misc: Ensure that the 'size' argument of PIL's 'resize' method is a tuple.
In Pillow 2.5.3, the code for the 'resize' function has a statement of the
form 'if self.size == size: ...'.  The expectation is that 'self.size' and
'size' are tuples.  If 'size' is a numpy array, the comparison raises
'ValueError: The truth value of an array with more than one element is
ambiguous.'

This commit ensures that the 'size' argument passed to the 'resize' method
is a tuple.

Closes gh-4032.