BUG: integrate: Fix handling of banded Jacobian by odeint.
The problem is that the Fortran jacobian callback function is passed a
Fortran array with dimension (2*ml+mu+1) x (n), but only the first
ml + mu + 1 rows are to be filled in by the user.  The extra ml rows
must be accounted for when the data returned from the call to the Python
callback are copied to the Fortran array.

(The problem fixed here has been fixed in a different way for the solvers
used in the 'ode' classes.)

Several changes:
o Add multipack_jac_type to the multipack global variables.  This integer
  has the value 1 (for the full Jacobian case) or 4 (for the banded Jacobian
  case).
o Replace the macro MATRIXC2F with the function 'copy_array_to_fortran' that
  can handle the destination array having a leading dimension that is not the
  same as the number of rows.
o Use 'copy_array_to_fortran' in the callback function 'ode_jacobian_function'.
  This allows the array returned by the user's jacobian function to be copied
  to the Fortran array without additional padding of the array generated by
  the user.
o An extension module written in Fortran and wrapped using f2py is now
  included in the 'tests' directory.  This module provides functions for
  evaluating the right-hand-side and Jacobian of a linear system with a
  banded Jacobian.  It also includes a function that solves the system
  by calling LSODA in Fortran.  To test that odeint has properly wrapped
  a system with a banded Jacobian, the "pure Fortran" use of LSODA can be
  compared to using odeint with Python wrappers of the right-hand-side
  function and Jacobian functions defined in the Fortran module.