BUG: sparse: Fix bench_sparse.py
In a previous change to bench_sparse.py, the imports of dok_matrix and
lil_matrix were removed, because pyflakes reported them as unused.
In fact, these were needed for a benchmark that was creating instances of
these matrices using eval().

In this commit, instead of constructing the names as a string and then
eval'ing the string, the class names are used explicitly.