ENH: vectorize some distance functions
as these are reductions they can't be autovectorized by gcc without
adding unsafe-math optimization flags, so vectorize them manually for
SSE2 which is available on all amd64 machines.
This changes the numerical result as the summation ordering is different,
but the change should be only in the order of a few machine precisions.