optimize: change df-sane eta_strategy once again
Restore the eta_strategy to eta = ||F0||^2 / (1 + k)^2.  This is similar
to the ||F0|| / (1 + k)^2 one in the DF-SANE paper, but replacing the
norm with its square to make it scale invariant.  For problems with
norms of order 1, the strategies are the same.

Based on test problems, although in principle the line search is what
ensures the iteration converges globally somewhere, in practice it
appears that the less it is needed the better. For problem F4, basing
the slack on the initial residual appears to outperform (convergence vs.
no convergence) using the current residual.  Looking at another problem,
it appears that using current residual can lead to excessive line
searching in late convergence stage.

Also fix up some style issues.