ENH: Make better use of the numpy API
A copy of the array is now created explicitly at the beginning of
procrustes and with these a few operations were improved. In addition
the readability of the check for zero-division was improved.

Additionally the readability of the data validation was improved through a
better use of the numpy API.