BUG: restore pseudobroading to lfilter, correct zf shape when broadcasting
lfilter requires that zi.ndim match x.ndim, and will not prepend
1s to acheive this.  It does allow dimensions of length 1, which
are expanded in the usual way.  It previously failed by always
forcing zf to have the same shape as zi, meaning it would generally
fail to return all of the final conditions if this broadcasting
behavior was used.  This commit corrects the shape of zf and
restores this pseudobroadcasting behavior that was removed while
fixing another bug.