BUG: Fixes #4408: Vector-valued constraints in minimize() et al
Using optimize.fmin_cobyla or optimize.minimize with method 'Cobyla' with vector-valued constraints used to throw a ValueError "setting an array element with a sequence" at line 240 in cobyla.py. The constraints functions could only return numbers for Cobyla, while SLSQP was able to take in vector constraints. Now Cobyla can take either a single constraint or a sequence of constraints, where each constraint function can return either a number or a sequence of numbers. This is true for both fmin_cobyla and minimize.