Accelerate major axis sequence slicing for compressed sparse matrices.
Indexing a CSR matrix with a slice such as X[[1, 100, 105], :] is
between 2x and 75x faster, with the greater performance improvements
coming when the size of the indexing list is small. This performance
improvements also apply to slicing a CSC matrix with a sequence of
columns.

Tests were added to ensure correct sequence slicing behavior.