Add tests showing problems with lsim
The implementation of lsim has a few problems, as mentioned in a number
of issues (e.g., #4071, #2652, #1104)

There are two causes for these problems.  One is that the current
implementation requires the computation of inv(A) (where the dynamics
are x' = A x + B u).  So for any system with a pole at the origin (e.g.,
an integrator), the algorithm will fail with 'LinAlgError: singular
matrix'.

The other problem is for systems with repeated eigenvalues (e.g.,
critically damped poles).  The existing routine computes the matrix
exponential using diagonalization, so if the A matrix is not
diagonalizable, the algorithm will give erroneous results.  This error
is more insidious, because an error message is not given; the routine
just gives wrong results, as pointed out in issue #2652 (and #4071).

This commit adds some tests that illustrate these problems.