Rework lsim; fixes #2652 and #1104
The previous implementation of lsim did not work for systems with a pole
at the origin, or with repeated poles.  This new version fixes all of
the (known) issues with lsim.

The new implementation uses the matrix exponential, like the old
implementation, but computes it using linalg.expm instead of by
diagonalization.  This solves the problem with repeated poles.  In
addition, the new implementation does not require inv(A) to compute the
forcing term, and it no longer fails for systems with a pole at the
origin.  Details of the algorithm are documented in the comments.

Note that there is already an lsim2 routine that fixes the above
problems.  However, it calls a generic ODE integrator, and can be
hundreds of times slower than the implementation here.

An important change with the new implementation is that the time vector
must now be evenly spaced.  Note that this same requirement is in the
MATLAB version of lsim.  One could modify this to allow for non-uniform
time samples, but this would slow down the algorithm, because the matrix
exponential would need to be recomputed whenever the timestep changes.
In addition, small, machine-precision-level differences in timestep
might trigger the matrix exponential to be recomputed when it is not
necessary.  Therefore, the implementation here checks whether timesteps
are approximately uniform using np.allclose(), and gives an error if
they are not.