BUG: sparse: ensure index dtype is large enough to pass all parameters to sparsetools
The index dtype size selection for CSR, CSC, BSR did not take into
account that the matrix size is also passed on to sparsetools, as
integers of the index dtype size.

Fix this by changing the dtype selection to account for also integer
parameters passed on to sparsetools.

Because the integer type is never downgraded, and the array shape
parameters do not change (outside .reshape which is unimplemented for
these types), it is sufficient to ensure this only in __init__