BUG: Fixes #4594: scipy.optimize.linprog IndexError when a callback is provided
The bug appeared at the beginning of Phase 2, if Phase 1 successfully terminated with some of the artificial variables as pivots. These artificial variables were zero because the corresponding element in the last column of T was zero, which allowed all artificial variables to be eliminated and Phase 1 to terminate successfully. However, basis[:m] contained elements that were greater than the length of solution, which resulted in IndexError at solution[basis[:m]]. The remaining true (non-artificial) variables that are not included in basis[:m] are also zero. Therefore simply setting the length of solution to be large enough solves the problem.