Extrapolate Anderson-Darling p-values linearly.
Currently, for out-of-range values of the Anderson-Darling k-sample
statistic, the *p*-value is computed by extrapolation by a quadratic
polynomial.  The original paper (Scholz and Stephens, 1987) suggests
linear extrapolation (see Sec. 7).  Such a change would give very
similar results for values of the statistic between the limits and
avoids having very large (>>1) extrapolated *p*-values when the
statistic is huge (i.e. beyond the minimum of the polynomial).