ENH: scipy.signal - Addition of spectrogram function
Given #4682, a spectrogram function is effectively free; it's just welch
without the averaging. I've added one test to make sure that the average of
the spectrogram segments agrees with the output of welch. Should more
tests be written, given that the only code difference between welch and
spectrogram is the averaging?