BUG: Fixes #4746: linprog returns solution violating equality constraint
This adds to pull request #4685. The problem happens if some artificial variable(s) remain basic at the end of Phase 1. PR #4685 solved the problem if the row in T associated with the artificial variable was zero (for all columns in T that correspond to a non-artificial variable). When there are non-zero elements in that row, then one of the non-zero elements is picked for pivot, and the artificial variable is replaced in the basis by the non-artificial variable. This is done until there are no more artificial variables in the basis that have non-zero rows in T.