BUG: spatial: mahalanobis edge case in pdist and cdist
In pdist and cdist with the Mahalanobis metric, when the covariance matrix or
its inverse are not provided, it is computed from the given observations.
For the covariance matrix to be nonsingular, the number of observations
must be more than the dimension of the observations.  Before this change, this
conditions was not checked.  The function might have raised a LinAlgError if
the inv() function detected the noninvertibility.  If the noninvertibility was
not detected, the inverse computed by np.linalg.inv()--and therefore the values
returned by the function--would be garbage.

With this change, a ValueError is raised when there are too few observations
given to pdist or cdist.