Ordering matters
After testing it seemed that it still wasn't working--annoyingly it appears that the order of the commands to matplotlib matters, as some formatting commands will overwrite/erase others.

Also, I added the resizing code to the other orientations. They're currently all based on the number of labels but that could be changed to the "p" parameter easily. This way seems simpler though.